requirements.txt
requests
bs4
lxml
selenium 
msedge-selenium-tools